@@ -1,5 +1,5 @@
 import {ChangeDetectionStrategy, Component, Inject, OnInit} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import {BoardProject} from '../../../model/board/data/project/project.model';
 import {BoardIssueView} from '../../../view-model/board/board-issue-view';
 import {Store} from '@ngrx/store';
