@@ -1,5 +1,5 @@
 import {ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {MatDialog} from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import {ParallelTask, ParallelTaskOption, ParallelTaskPosition} from '../../../model/board/data/project/project.model';
 import {ParallelTaskSelectorComponent} from './parallel-task-selector.component';
 import {BoardIssueView} from '../../../view-model/board/board-issue-view';
