@@ -1,7 +1,7 @@
 /**
  * Abstract base class for a board containing a fixed header.
  */
-import {ChangeDetectorRef, EventEmitter, Input, NgZone, OnChanges, Output, SimpleChange, SimpleChanges} from '@angular/core';
+import { ChangeDetectorRef, EventEmitter, Input, NgZone, OnChanges, Output, SimpleChange, SimpleChanges, Directive } from '@angular/core';
 import {BoardViewModel} from '../../../view-model/board/board-view';
 import {BOARD_HEADERS_HEIGHT, TOOLBAR_HEIGHT} from '../../../common/view-constants';
 import {IssueDetailState} from '../../../model/board/user/issue-detail/issue-detail.model';
@@ -9,6 +9,7 @@ import {BehaviorSubject, Subject, Observable} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 import {BoardHeader} from '../../../view-model/board/board-header';
 
+@Directive()
 export class FixedHeaderView implements OnChanges {
 
   @Input()
